+++ Sources/SwiftGenKit/Parsers/Strings/StringsEntry.swift
@@ -12,19 +12,21 @@ extension Strings {
     var comment: String?
     let key: String
     let translation: String
+    let formatKey: String?
     let types: [PlaceholderType]
     let keyStructure: [String]
 
-    init(key: String, translation: String, types: [PlaceholderType], keyStructureSeparator: String) {
+    init(key: String, formatKey: String? = nil, translation: String, types: [PlaceholderType], keyStructureSeparator: String) {
       self.key = key
       self.translation = translation
+      self.formatKey = formatKey
       self.types = types
       self.keyStructure = Self.split(key: key, separator: keyStructureSeparator)
     }
 
-    init(key: String, translation: String, keyStructureSeparator: String) throws {
+    init(key: String, formatKey: String? = nil, translation: String, keyStructureSeparator: String) throws {
       let types = try PlaceholderType.placeholderTypes(fromFormat: translation)
-      self.init(key: key, translation: translation, types: types, keyStructureSeparator: keyStructureSeparator)
+      self.init(key: key, formatKey: formatKey, translation: translation, types: types, keyStructureSeparator: keyStructureSeparator)
     }
 
     // MARK: - Structured keys